rearrange functions to get close to the original
