+++ isle.mak
@@ -299,7 +299,6 @@ CLEAN :
 	-@erase "$(INTDIR)\define.obj"
 	-@erase "$(INTDIR)\isle.obj"
 	-@erase "$(INTDIR)\isle.res"
-	-@erase "$(INTDIR)\main.obj"
 	-@erase "$(INTDIR)\vc40.pdb"
 	-@erase ".\Release\ISLE.EXE"
 	-@erase ".\Release\ISLE.PDB"
@@ -360,7 +359,6 @@ LINK32_OBJS= \
 	"$(INTDIR)\define.obj" \
 	"$(INTDIR)\isle.obj" \
 	"$(INTDIR)\isle.res" \
-	"$(INTDIR)\main.obj" \
 	".\Release\LEGO1.LIB"
 
 ".\Release\ISLE.EXE" : "$(OUTDIR)" $(DEF_FILE) $(LINK32_OBJS)
@@ -389,7 +387,6 @@ CLEAN :
 	-@erase "$(INTDIR)\define.obj"
 	-@erase "$(INTDIR)\isle.obj"
 	-@erase "$(INTDIR)\isle.res"
-	-@erase "$(INTDIR)\main.obj"
 	-@erase "$(INTDIR)\vc40.idb"
 	-@erase "$(INTDIR)\vc40.pdb"
 	-@erase ".\Debug\ISLE.EXE"
@@ -452,7 +449,6 @@ LINK32_OBJS= \
 	"$(INTDIR)\define.obj" \
 	"$(INTDIR)\isle.obj" \
 	"$(INTDIR)\isle.res" \
-	"$(INTDIR)\main.obj" \
 	".\LEGO1\Debug\LEGO1.lib"
 
 ".\Debug\ISLE.EXE" : "$(OUTDIR)" $(DEF_FILE) $(LINK32_OBJS)
@@ -902,20 +898,10 @@ DEP_CPP_MXPAL=\
 ################################################################################
 # Begin Source File
 
-SOURCE=.\ISLE\define.cpp
-DEP_CPP_DEFIN=\
-	".\ISLE\define.h"\
-	
-
-"$(INTDIR)\define.obj" : $(SOURCE) $(DEP_CPP_DEFIN) "$(INTDIR)"
-   $(CPP) $(CPP_PROJ) $(SOURCE)
-
-
-# End Source File
-################################################################################
-# Begin Source File
-
 SOURCE=.\ISLE\isle.cpp
+
+!IF  "$(CFG)" == "ISLE - Win32 Release"
+
 DEP_CPP_ISLE_=\
 	".\ISLE\define.h"\
 	".\ISLE\isle.h"\
@@ -974,29 +960,33 @@ DEP_CPP_ISLE_=\
    $(CPP) $(CPP_PROJ) $(SOURCE)
 
 
-# End Source File
-################################################################################
-# Begin Source File
+!ELSEIF  "$(CFG)" == "ISLE - Win32 Debug"
 
-SOURCE=.\ISLE\main.cpp
-DEP_CPP_MAIN_=\
+DEP_CPP_ISLE_=\
 	".\ISLE\define.h"\
 	".\ISLE\isle.h"\
+	".\ISLE\res\resource.h"\
 	".\LEGO1\lego3dmanager.h"\
 	".\LEGO1\lego3dview.h"\
+	".\LEGO1\legoanimationmanager.h"\
+	".\LEGO1\legobuildingmanager.h"\
 	".\LEGO1\legoentity.h"\
 	".\LEGO1\legogamestate.h"\
 	".\LEGO1\legoinc.h"\
 	".\LEGO1\legoinputmanager.h"\
+	".\LEGO1\legomodelpresenter.h"\
 	".\LEGO1\legonavcontroller.h"\
 	".\LEGO1\legoomni.h"\
+	".\LEGO1\legopartpresenter.h"\
 	".\LEGO1\legoroi.h"\
 	".\LEGO1\legovideomanager.h"\
+	".\LEGO1\legoworldpresenter.h"\
 	".\LEGO1\mxatomid.h"\
 	".\LEGO1\mxbackgroundaudiomanager.h"\
 	".\LEGO1\mxbool.h"\
 	".\LEGO1\mxcore.h"\
 	".\LEGO1\mxcriticalsection.h"\
+	".\LEGO1\mxdirectdraw.h"\
 	".\LEGO1\mxdsaction.h"\
 	".\LEGO1\mxdsfile.h"\
 	".\LEGO1\mxdsobject.h"\
@@ -1026,9 +1016,11 @@ DEP_CPP_MAIN_=\
 	".\LEGO1\viewmanager.h"\
 	
 
-"$(INTDIR)\main.obj" : $(SOURCE) $(DEP_CPP_MAIN_) "$(INTDIR)"
+"$(INTDIR)\isle.obj" : $(SOURCE) $(DEP_CPP_ISLE_) "$(INTDIR)"
    $(CPP) $(CPP_PROJ) $(SOURCE)
 
+
+!ENDIF 
 
 # End Source File
 ################################################################################
@@ -1168,18 +1160,6 @@ SOURCE=.\LEGO1\legovideomanager.h
 # Begin Source File
 
 SOURCE=.\LEGO1\legoworldpresenter.h
-
-!IF  "$(CFG)" == "ISLE - Win32 Release"
-
-!ELSEIF  "$(CFG)" == "ISLE - Win32 Debug"
-
-!ENDIF 
-
-# End Source File
-################################################################################
-# Begin Source File
-
-SOURCE=.\LEGO1\mxatomid.h
 
 !IF  "$(CFG)" == "ISLE - Win32 Release"
 
@@ -1470,6 +1450,46 @@ SOURCE=.\LEGO1\mxvideoparamflags.h
 !ENDIF 
 
 # End Project Dependency
+################################################################################
+# Begin Source File
+
+SOURCE=.\LEGO1\mxatomid.h
+
+!IF  "$(CFG)" == "ISLE - Win32 Release"
+
+!ELSEIF  "$(CFG)" == "ISLE - Win32 Debug"
+
+!ENDIF 
+
+# End Source File
+################################################################################
+# Begin Source File
+
+SOURCE=.\ISLE\define.cpp
+
+!IF  "$(CFG)" == "ISLE - Win32 Release"
+
+DEP_CPP_DEFIN=\
+	".\ISLE\define.h"\
+	
+
+"$(INTDIR)\define.obj" : $(SOURCE) $(DEP_CPP_DEFIN) "$(INTDIR)"
+   $(CPP) $(CPP_PROJ) $(SOURCE)
+
+
+!ELSEIF  "$(CFG)" == "ISLE - Win32 Debug"
+
+DEP_CPP_DEFIN=\
+	".\ISLE\define.h"\
+	
+
+"$(INTDIR)\define.obj" : $(SOURCE) $(DEP_CPP_DEFIN) "$(INTDIR)"
+   $(CPP) $(CPP_PROJ) $(SOURCE)
+
+
+!ENDIF 
+
+# End Source File
 # End Target
 # End Project
 ################################################################################
